Improvments to dconf defaults handling
A new variable: gnome_aur_helper will now configure the helper
program (if any). This will be used with the default settings for the
arch-update extension in favor of the yaourt stat task. The result is
support for more helpers and one less task.

Also remove default deja-dup path, which does make a few too many
assumptions for a reusable role. It never seemed to work anyways.