Add support for aur helper - pacaur
In terms of usability pacaur seems the closest to yaourt so it is
easy to add.

A lot of helpers have different syntax, even if it is "Pacman-like".
Therefore there can no longer be a generic helper syntax (-Syua), but
instead each helper needs to be added to the template.